moved unit test for RoslynCodeTaskFactory
